game: adding Dijkstra maps in navigation.hpp
Inspired by:
<https://chizaruu.github.io/roguebasin/the_incredible_power_of_dijkstra_maps>

- Adding Dijkstra map: tile updating and map balancing
- LevelWrapper::set will not update dijkstra maps. (they use it)
- Spawning Hero outside of level. (causes recursion issues otherwise)
- Updating full map at game start
- Level::unhash_chunk_pos is useful for iterating chunks
- Minimum zoom 0.5 -> 0.3
- Adding interest dijkstra map
- Adding tick_hero()
- bool operator <(Point, Point) needed to use std::set<Point>
- Fixing warnings

As always, these are not tested in advance.

+++ src/game/render_tile.cpp
@@ -55,8 +55,8 @@ void game::render_tile(int x, int y)
     };
 
     SDL_FRect dest {
-        x * 32,
-        y * 32,
+        static_cast<float>(x * 32),
+        static_cast<float>(y * 32),
         32, 32
     };
